+++ src/app/(home)/page.tsx
@@ -1,10 +1,9 @@
-import { InputForm } from "@/components/ui/FormInput";
 import Image from "next/image";
 
 export default function Home() {
   return (
     <>
-      <InputForm />
+      <h1>oi</h1>
     </>
   );
 }
