+++ src/app/form/fotos/page.tsx
@@ -1,0 +1,97 @@
+"use client";
+
+import React from "react";
+import {
+  Form,
+  FormControl,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@/components/ui/form";
+
+import { ToastContainer, toast } from "react-toastify";
+import "react-toastify/dist/ReactToastify.css";
+
+import { z } from "zod";
+import { zodResolver } from "@hookform/resolvers/zod";
+
+import { Input } from "@/components/ui/input";
+
+import { useToast } from "@/components/ui/use-toast";
+
+import { useRouter } from "next/navigation";
+
+import { useForm } from "react-hook-form";
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardFooter,
+  CardHeader,
+  CardTitle,
+} from "@/components/ui/card";
+import { Button } from "@/components/ui/button";
+
+const FormSchema = z.object({
+  picture1: z.string().nonempty({
+    message: "Selecione uma opção valida.",
+  }),
+});
+
+export default function FormProblem() {
+  const router = useRouter();
+  const notifySucess = () => toast.success("Wow so easy!");
+
+  const form = useForm<z.infer<typeof FormSchema>>({
+    resolver: zodResolver(FormSchema),
+    defaultValues: {
+      picture1: "",
+    },
+  });
+
+  function onSubmit(data: z.infer<typeof FormSchema>) {
+    console.log(data);
+    notifySucess();
+  }
+  return (
+    <Card className="flex flex-col mt-9 m-3">
+      <CardHeader>
+        <CardTitle>Fotos e Documentação</CardTitle>
+        <CardDescription>
+          Preencha corretamente as questões abaixo!
+        </CardDescription>
+      </CardHeader>
+      <CardContent>
+        <Form {...form}>
+          <form onSubmit={form.handleSubmit(onSubmit)}>
+            <div className="flex justify-center flex-col m-3 gap-2">
+              <FormField
+                control={form.control}
+                name="picture1"
+                render={({ field }) => (
+                  <FormItem>
+                    <FormLabel>
+                      Por favor, anexe fotos do local afetado.
+                    </FormLabel>
+                    <FormControl>
+                      <Input id="picture" type="file" {...field} />
+                    </FormControl>
+                    <FormMessage />
+                  </FormItem>
+                )}
+              />
+            </div>
+
+            <CardFooter className="flex justify-between">
+              <Button variant="outline" onClick={() => router.back()}>
+                Back
+              </Button>
+              <Button type="submit">Next</Button>
+            </CardFooter>
+          </form>
+        </Form>
+      </CardContent>
+    </Card>
+  );
+}
