+++ src/app/form/infra/page.tsx
@@ -1,0 +1,149 @@
+"use client";
+import {
+  Form,
+  FormControl,
+  FormDescription,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@/components/ui/form";
+
+import { useRouter } from "next/navigation";
+
+import { z } from "zod";
+import { zodResolver } from "@hookform/resolvers/zod";
+
+import { Input } from "@/components/ui/input";
+
+import { toast } from "@/components/ui/use-toast";
+
+import {
+  Select,
+  SelectContent,
+  SelectGroup,
+  SelectItem,
+  SelectLabel,
+  SelectTrigger,
+  SelectValue,
+} from "@/components/ui/select";
+
+import { useForm } from "react-hook-form";
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardFooter,
+  CardHeader,
+  CardTitle,
+} from "@/components/ui/card";
+import { Button } from "@/components/ui/button";
+
+const FormSchema = z.object({
+  question1: z.string().min(3, {
+    message: "Digite pelo menos 3 caracteres.",
+  }),
+  selectField: z.string().nonempty({
+    message: "Selecione uma opção valida.",
+  }),
+});
+
+export default function FormInfra() {
+  const router = useRouter();
+  const form = useForm<z.infer<typeof FormSchema>>({
+    resolver: zodResolver(FormSchema),
+    defaultValues: {
+      selectField: "",
+      question1: "",
+    },
+  });
+
+  function onSubmit(data: z.infer<typeof FormSchema>) {
+    console.log(JSON.stringify(data).length);
+
+    toast({
+      title: "You submitted the following values:",
+      description: (
+        <pre className="mt-2 w-[340px] rounded-md bg-slate-950 p-4">
+          <code className="text-white">{JSON.stringify(data, null, 2)}</code>
+        </pre>
+      ),
+    });
+    router.push("/form/problems");
+  }
+  return (
+    <Card className="flex flex-col mt-9 m-3">
+      <CardHeader>
+        <CardTitle>Tipo de Infraestrutura</CardTitle>
+        <CardDescription>
+          Preencha corretamente as questões abaixo!
+        </CardDescription>
+      </CardHeader>
+      <CardContent>
+        <Form {...form}>
+          <form onSubmit={form.handleSubmit(onSubmit)}>
+            <div className="flex justify-center flex-col m-3 gap-2">
+              <FormField
+                control={form.control}
+                name="selectField"
+                render={({ field }) => (
+                  <FormItem>
+                    <FormLabel>
+                      Qual tipo de infraestrutura foi afetada?
+                    </FormLabel>
+                    <FormControl>
+                      <Select
+                        onValueChange={field.onChange}
+                        value={field.value}
+                      >
+                        <SelectTrigger className="w-[180px]">
+                          <SelectValue placeholder="Selecione uma opção" />
+                        </SelectTrigger>
+                        <SelectContent>
+                          <SelectGroup>
+                            <SelectLabel>Infraestrutura</SelectLabel>
+                            <SelectItem value="school">Escola</SelectItem>
+                            <SelectItem value="healthCenter">
+                              Posto de saúde
+                            </SelectItem>
+                            <SelectItem value="road">Estrada</SelectItem>
+                            <SelectItem value="bridge">Ponte</SelectItem>
+                            <SelectItem value="sanitation">
+                              Rede de saneamento
+                            </SelectItem>
+                            <SelectItem value="other">
+                              Outro (especificar)
+                            </SelectItem>
+                          </SelectGroup>
+                        </SelectContent>
+                      </Select>
+                    </FormControl>
+                    <FormMessage />
+                  </FormItem>
+                )}
+              />
+              <FormField
+                control={form.control}
+                name="question1"
+                render={({ field }) => (
+                  <FormItem>
+                    <FormLabel>
+                      Qual é o nome/identificação da infraestrutura?
+                    </FormLabel>
+                    <FormControl>
+                      <Input placeholder="Digite aqui..." {...field} />
+                    </FormControl>
+                    <FormMessage />
+                  </FormItem>
+                )}
+              />
+            </div>
+            <CardFooter className="flex justify-between">
+              <Button type="submit">Next</Button>
+            </CardFooter>
+          </form>
+        </Form>
+      </CardContent>
+    </Card>
+  );
+}
